test_task/test_task/requirements.txt
selenium==4.14.0
pytest==7.4.3
allure-pytest==2.13.2
